iamfastapi/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose
passlib
argon2-cffi
email-validator
pydantic[email]
